Revert "fix circular import error"
This reverts commit 17c2f925a8aa46510f0c45aa2e55092c29b45117.
